Merge to Fronted

+++ app/Controllers/Admin/Products.php
@@ -99,11 +99,9 @@ class Products extends BaseController
                 $iconOld = $this->db->table($this->table)->select('icon')->where([EncKey('id') => Input_('id')])->get()->getRow()->icon;
 
                 // jika cover ada maka hapus filenya
-                if ($iconOld != '') {
-                    $path = ROOTPATH.'public/uploads/products/'.$iconOld;
-                    if (file_exists($path)) {
-                        unlink($path);
-                    }
+                if ($iconOld != "") {
+                    $path = FILESDIR . '/uploads/products/' . $iconOld;
+                    if (file_exists($path)) unlink($path);
                 }
 
                 // upload cover baru
@@ -302,8 +300,8 @@ class Products extends BaseController
                 throw new \Exception($this->validator->listErrors());
             }
             $file = $this->request->getFile('icon');
-            $fileName = time().'_'.$file->getName();
-            $path = ROOTPATH.'public/uploads/products/';
+            $fileName = time() . "_" . $file->getName();
+            $path = FILESDIR . '/uploads/products/';
             $file->move($path, $fileName);
             $result = Update($this->table, ['icon' => $fileName], $idArticle);
         } catch (\Throwable $th) {
