+++ app/Controllers/Admin/Profile.php
@@ -14,14 +14,18 @@ class Profile extends BaseController
 
     public function index()
     {
+        // echo __DIR__ . "<br>";
+        // echo FILESDIR;
+        // Print_(FILESDIR, false, false);
+        // exit();
         $data = [
             'title' => 'Profile',
             'menu' => 'profile',
             'roti' => [
                 'Home:blank' => base_url(),
-                'Dashboard' => base_url(ADMIN_PATH.'/dashboard'),
+                'Dashboard' => base_url(ADMIN_PATH . '/dashboard'),
                 'Profile' => '',
-                session('name').':active' => '',
+                session('name') . ':active' => '',
             ],
         ];
 
@@ -53,11 +57,10 @@ class Profile extends BaseController
                 $oldPhoto = $this->db->table($this->table)->select('photo')->where(['id' => session('userId')])->get()->getRow()->photo;
 
                 // jika phptp ada maka hapus filenya
-                if ($oldPhoto != '') {
-                    $path = ROOTPATH.'public/uploads/users/'.$oldPhoto;
-                    if (file_exists($path)) {
-                        unlink($path);
-                    }
+                if ($oldPhoto != "") {
+                    // $path = ROOTPATH . 'public/uploads/users/' . $oldPhoto;
+                    $path = FILESDIR . '/uploads/users/' . $oldPhoto;
+                    if (file_exists($path)) unlink($path);
                 }
                 // upload photo baru
                 $updatePhoto = $this->uploadPhoto();
@@ -104,9 +107,10 @@ class Profile extends BaseController
                 throw new \Exception($this->validator->listErrors());
             }
             $file = $this->request->getFile('photo');
-            $fileName = time().'_'.$file->getName();
+            $fileName = time() . '_' . $file->getName();
             session()->set('photo', $fileName);
-            $path = ROOTPATH.'public_html/uploads/users/';
+            // $path = ROOTPATH . 'public/uploads/users/';
+            $path = FILESDIR . '/uploads/users/';
             $file->move($path, $fileName);
             $result = Update($this->table, ['photo' => $fileName], ['id' => session('userId')]);
         } catch (\Throwable $th) {
@@ -127,8 +131,6 @@ class Profile extends BaseController
     public function setPassword()
     {
         try {
-            //code...
-
             $validate = Validate([
                 'password' => 'required|password',
                 'passwordLama' => 'required',
@@ -345,8 +347,8 @@ class Profile extends BaseController
 
     public function previewWeb()
     {
-        echo '<style>'.$_REQUEST['css'].'</style>';
+        echo '<style>' . $_REQUEST['css'] . '</style>';
         echo $_REQUEST['html'];
-        echo '<script>'.$_REQUEST['js'].'</script>';
+        echo '<script>' . $_REQUEST['js'] . '</script>';
     }
 }
